Committing the java interface to the native stuff.  Just have to finish the CPP part.

+++ DarpaRoboticsChallenge/src/us/ihmc/darpaRoboticsChallenge/ros/messages/Float64Message.java
@@ -8,7 +8,7 @@ public class Float64Message
 
    private double value;
 
-   public void setValue(float value)
+   public void setValue(double value)
    {
       this.value = value;
    }
